+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/engine300_300/ObjectFactory.java
@@ -100,14 +100,6 @@ public class ObjectFactory {
   }
 
   /**
-   * Create an instance of {@link RelationshipEndTranslation.Annotations }
-   *
-   */
-  public RelationshipEndTranslation.Annotations createRelationshipEndTranslationAnnotations() {
-    return new RelationshipEndTranslation.Annotations();
-  }
-
-  /**
    * Create an instance of {@link RelationshipEnd.Translations }
    *
    */
