+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/AggregationAttribute.java
@@ -13,8 +13,6 @@
 */
 package org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla;
 
-import java.util.List;
-
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
 import jakarta.xml.bind.annotation.XmlElement;
@@ -29,7 +27,7 @@ public class AggregationAttribute {
   @XmlElement(name = "AttributeID", required = true)
   protected String attributeID;
   @XmlElement(name = "Annotations")
-  protected AggregationAttribute.Annotations annotations;
+  protected Annotations annotations;
 
   public String getAttributeID() {
     return attributeID;
@@ -39,28 +37,12 @@ public class AggregationAttribute {
     this.attributeID = value;
   }
 
-  public AggregationAttribute.Annotations getAnnotations() {
+  public Annotations getAnnotations() {
     return annotations;
   }
 
-  public void setAnnotations(AggregationAttribute.Annotations value) {
+  public void setAnnotations(Annotations value) {
     this.annotations = value;
-  }
-
-  @XmlAccessorType(XmlAccessType.FIELD)
-  @XmlType(name = "", propOrder = { "annotation" })
-  public static class Annotations {
-
-    @XmlElement(name = "Annotation")
-    protected List<Annotation> annotation;
-
-    public List<Annotation> getAnnotation() {
-      return this.annotation;
-    }
-
-      public void setAnnotation(List<Annotation> annotation) {
-          this.annotation = annotation;
-      }
   }
 
 }
