+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/AttributeTranslation.java
@@ -13,8 +13,6 @@
  */
 package org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla;
 
-import java.util.List;
-
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
 import jakarta.xml.bind.annotation.XmlElement;
@@ -37,7 +35,7 @@ public class AttributeTranslation {
     @XmlElement(name = "DisplayFolder")
     protected String displayFolder;
     @XmlElement(name = "Annotations")
-    protected AttributeTranslation.Annotations annotations;
+    protected Annotations annotations;
     @XmlElement(name = "CaptionColumn")
     protected DataItem captionColumn;
     @XmlElement(name = "MembersWithDataCaption")
@@ -75,11 +73,11 @@ public class AttributeTranslation {
         this.displayFolder = value;
     }
 
-    public AttributeTranslation.Annotations getAnnotations() {
+    public Annotations getAnnotations() {
         return annotations;
     }
 
-    public void setAnnotations(AttributeTranslation.Annotations value) {
+    public void setAnnotations(Annotations value) {
         this.annotations = value;
     }
 
@@ -97,22 +95,6 @@ public class AttributeTranslation {
 
     public void setMembersWithDataCaption(String value) {
         this.membersWithDataCaption = value;
-    }
-
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {"annotation"})
-    public static class Annotations {
-
-        @XmlElement(name = "Annotation")
-        protected List<Annotation> annotation;
-
-        public List<Annotation> getAnnotation() {
-            return this.annotation;
-        }
-
-        public void setAnnotation(List<Annotation> annotation) {
-            this.annotation = annotation;
-        }
     }
 
 }
