+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/CubeDimensionPermission.java
@@ -37,7 +37,7 @@ public class CubeDimensionPermission {
     @XmlElement(name = "AttributePermissions")
     protected CubeDimensionPermission.AttributePermissions attributePermissions;
     @XmlElement(name = "Annotations")
-    protected CubeDimensionPermission.Annotations annotations;
+    protected Annotations annotations;
 
     public String getCubeDimensionID() {
         return cubeDimensionID;
@@ -79,28 +79,12 @@ public class CubeDimensionPermission {
         this.attributePermissions = value;
     }
 
-    public CubeDimensionPermission.Annotations getAnnotations() {
+    public Annotations getAnnotations() {
         return annotations;
     }
 
-    public void setAnnotations(CubeDimensionPermission.Annotations value) {
+    public void setAnnotations(Annotations value) {
         this.annotations = value;
-    }
-
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {"annotation"})
-    public static class Annotations {
-
-        @XmlElement(name = "Annotation")
-        protected List<Annotation> annotation;
-
-        public List<Annotation> getAnnotation() {
-            return this.annotation;
-        }
-
-        public void setAnnotation(List<Annotation> annotation) {
-            this.annotation = annotation;
-        }
     }
 
     @XmlAccessorType(XmlAccessType.FIELD)
