+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/CubeHierarchy.java
@@ -13,8 +13,6 @@
  */
 package org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla;
 
-import java.util.List;
-
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
 import jakarta.xml.bind.annotation.XmlElement;
@@ -35,7 +33,7 @@ public class CubeHierarchy {
     @XmlElement(name = "Enabled")
     protected Boolean enabled;
     @XmlElement(name = "Annotations")
-    protected CubeHierarchy.Annotations annotations;
+    protected Annotations annotations;
 
     public String getHierarchyID() {
         return hierarchyID;
@@ -69,27 +67,12 @@ public class CubeHierarchy {
         this.enabled = value;
     }
 
-    public CubeHierarchy.Annotations getAnnotations() {
+    public Annotations getAnnotations() {
         return annotations;
     }
 
-    public void setAnnotations(CubeHierarchy.Annotations value) {
+    public void setAnnotations(Annotations value) {
         this.annotations = value;
     }
 
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {"annotation"})
-    public static class Annotations {
-
-        @XmlElement(name = "Annotation")
-        protected List<Annotation> annotation;
-
-        public List<Annotation> getAnnotation() {
-            return this.annotation;
-        }
-
-        public void setAnnotation(List<Annotation> annotation) {
-            this.annotation = annotation;
-        }
-    }
 }
