+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/Measure.java
@@ -59,7 +59,7 @@ public class Measure {
     @XmlElement(name = "Translations")
     protected Measure.Translations translations;
     @XmlElement(name = "Annotations")
-    protected Measure.Annotations annotations;
+    protected Annotations annotations;
 
     public String getName() {
         return name;
@@ -189,28 +189,12 @@ public class Measure {
         this.translations = value;
     }
 
-    public Measure.Annotations getAnnotations() {
+    public Annotations getAnnotations() {
         return annotations;
     }
 
-    public void setAnnotations(Measure.Annotations value) {
+    public void setAnnotations(Annotations value) {
         this.annotations = value;
-    }
-
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {"annotation"})
-    public static class Annotations {
-
-        @XmlElement(name = "Annotation")
-        protected List<Annotation> annotation;
-
-        public List<Annotation> getAnnotation() {
-            return this.annotation;
-        }
-
-        public void setAnnotation(List<Annotation> annotation) {
-            this.annotation = annotation;
-        }
     }
 
     @XmlAccessorType(XmlAccessType.FIELD)
