+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/MeasureGroupAttribute.java
@@ -33,7 +33,7 @@ public class MeasureGroupAttribute {
     @XmlElement(name = "Type")
     protected String type;
     @XmlElement(name = "Annotations")
-    protected MeasureGroupAttribute.Annotations annotations;
+    protected Annotations annotations;
 
     public String getAttributeID() {
         return attributeID;
@@ -59,29 +59,14 @@ public class MeasureGroupAttribute {
         this.type = value;
     }
 
-    public MeasureGroupAttribute.Annotations getAnnotations() {
+    public Annotations getAnnotations() {
         return annotations;
     }
 
-    public void setAnnotations(MeasureGroupAttribute.Annotations value) {
+    public void setAnnotations(Annotations value) {
         this.annotations = value;
     }
 
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {"annotation"})
-    public static class Annotations {
-
-        @XmlElement(name = "Annotation")
-        protected List<Annotation> annotation;
-
-        public List<Annotation> getAnnotation() {
-            return this.annotation;
-        }
-
-        public void setAnnotation(List<Annotation> annotation) {
-            this.annotation = annotation;
-        }
-    }
 
     @XmlAccessorType(XmlAccessType.FIELD)
     @XmlType(name = "", propOrder = {"keyColumn"})
