+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/PerspectiveHierarchy.java
@@ -13,8 +13,6 @@
  */
 package org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla;
 
-import java.util.List;
-
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
 import jakarta.xml.bind.annotation.XmlElement;
@@ -29,7 +27,7 @@ public class PerspectiveHierarchy {
     @XmlElement(name = "HierarchyID", required = true)
     protected String hierarchyID;
     @XmlElement(name = "Annotations")
-    protected PerspectiveHierarchy.Annotations annotations;
+    protected Annotations annotations;
 
     public String getHierarchyID() {
         return hierarchyID;
@@ -39,28 +37,12 @@ public class PerspectiveHierarchy {
         this.hierarchyID = value;
     }
 
-    public PerspectiveHierarchy.Annotations getAnnotations() {
+    public Annotations getAnnotations() {
         return annotations;
     }
 
-    public void setAnnotations(PerspectiveHierarchy.Annotations value) {
+    public void setAnnotations(Annotations value) {
         this.annotations = value;
-    }
-
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {"annotation"})
-    public static class Annotations {
-
-        @XmlElement(name = "Annotation")
-        protected List<Annotation> annotation;
-
-        public List<Annotation> getAnnotation() {
-            return this.annotation;
-        }
-
-        public void setAnnotation(List<Annotation> annotation) {
-            this.annotation = annotation;
-        }
     }
 
 }
